networkd: do not assume routes are classless
RFC 3442 adds support for classless static routes to DHCP. When that
option is used classful static routes (option 33) and the router option
should be ignored. Currently networkd assumes all static routes (both
option 33 and option 121) are classless and always ignores the gateway
route if either are specified. Furthermore it currently concatenates the
classless routes to the classful routes, instead of ignoring the
classful ones when both are specified.

 - Keep track of which type of routes it is handling and do not assume
   they are classless.
 - Free then overwrite the classful routes when classless routes are
   specified and valid.
 - Log when classful routes are replaced by classless ones.
 - Do not ignore the router option when classless routes are not
   specified.
 - Handle the gateway route before handling static routes in case they
   need gateway route to be reachable and prevent 'Network is
   unreachable' errors.